shifted update query in query.php fom crud_post_controller
