+++ src/model/Query.php
@@ -26,6 +26,28 @@ function signUp($conn, $table, $first_name, $last_name, $user_name, $user_email,
     }
 }
 
+// Update Query
+function update($conn, $table_name, $set_data, $condition)
+{
+    try {
+        if (!empty($condition)) {
+            $update_query = "UPDATE $table_name SET $set_data WHERE $condition";
+
+            $update_result = mysqli_query($conn, $update_query);
+            if ($update_result) {
+                return $update_result;
+            } else {
+                return false;
+            }
+        } else {
+            echo "No Condition Available";
+            return false;
+        }
+    } catch (Exception $e) {
+        echo "Error : " . $e->getMessage();
+        return false;
+    }
+}
 
 // Function to let users log in 
 function signIn($conn, $table, $user_email, $user_password)
